// lab5/WycieczkiFirebase/Wycieczki/src/app/model/trip.ts
export interface Trip {
    id: number,
    name: string;
    country: string;
    startDate: string;
    endDate: string;
    price: number;
    capacity: number;
    amount: number;
    description: string;
    image: string;
    image1: string;
    image2: string;
    actualimage: string;
    likes: number;
    dislikes: number;
  }